Decouple SharkContext from SQLSuite. Put it in a SharkRunner singleton so that we can run SQL queries across different test suites.
